requirements.txt
python-telegram-bot[rate-limiter]==20.1
PyYAML==6.0
pytz
